+++ tw-tasks-core-spring-boot-starter/src/main/java/com/transferwise/tasks/core/autoconfigure/TwTasksCoreAutoConfiguration.java
@@ -4,9 +4,11 @@ import com.transferwise.common.baseutils.concurrency.DefaultExecutorServicesProv
 import com.transferwise.common.baseutils.concurrency.IExecutorServicesProvider;
 import com.transferwise.common.baseutils.transactionsmanagement.TransactionsConfiguration;
 import com.transferwise.common.gracefulshutdown.GracefulShutdowner;
+import com.transferwise.tasks.IMigrationHandler;
 import com.transferwise.tasks.IPriorityManager;
 import com.transferwise.tasks.ITaskDataSerializer;
 import com.transferwise.tasks.ITasksService;
+import com.transferwise.tasks.MigrationHandler;
 import com.transferwise.tasks.PriorityManager;
 import com.transferwise.tasks.TaskDataSerializer;
 import com.transferwise.tasks.TasksProperties;
@@ -163,7 +165,7 @@ public class TwTasksCoreAutoConfiguration {
 
   @Bean
   @ConditionalOnMissingBean(ITopicPartitionsManager.class)
-  public ITopicPartitionsManager twTasksTopicPartitionsManager() {
+  public NoOpTopicPartitionsManager twTasksTopicPartitionsManager() {
     return new NoOpTopicPartitionsManager();
   }
 
@@ -226,25 +228,31 @@ public class TwTasksCoreAutoConfiguration {
 
   @Bean
   @ConditionalOnMissingBean(IMdcService.class)
-  public IMdcService twTasksMdcService() {
+  public MdcService twTasksMdcService() {
     return new MdcService();
   }
 
   @Bean
   @ConditionalOnMissingBean(IEntryPointsService.class)
-  public IEntryPointsService twTasksEntryPointsService() {
+  public EntryPointsService twTasksEntryPointsService() {
     return new EntryPointsService();
   }
 
   @Bean
   @ConditionalOnMissingBean(ITaskDaoDataSerializer.class)
-  public ITaskDaoDataSerializer twTasksTaskDaoDataSerializer() {
+  public TaskDaoDataSerializer twTasksTaskDaoDataSerializer() {
     return new TaskDaoDataSerializer();
   }
 
   @Bean
   @ConditionalOnMissingBean(ITaskDataSerializer.class)
-  public ITaskDataSerializer taskDataSerializer() {
+  public TaskDataSerializer taskDataSerializer() {
     return new TaskDataSerializer();
+  }
+
+  @Bean
+  @ConditionalOnMissingBean(IMigrationHandler.class)
+  public MigrationHandler migrationHandler() {
+    return new MigrationHandler();
   }
 }