+++ tw-tasks-core/src/main/java/com/transferwise/tasks/dao/MySqlTaskDao.java
@@ -222,8 +222,12 @@ public class MySqlTaskDao implements ITaskDao {
       Connection con = DataSourceUtils.getConnection(dataSource);
       try {
         try (PreparedStatement ps = con.prepareStatement(insertTaskSql)) {
+          String data = "";
+          if (tasksProperties.getMigration().isCopyDataToTwTaskField() && request.getData() != null) {
+            data = new String(request.getData(), StandardCharsets.UTF_8);
+          }
           args(taskId, request.getType(), request.getSubType(),
-              request.getStatus(), "", nextEventTime, now, now, now, 0, 0, request.getPriority())
+              request.getStatus(), data, nextEventTime, now, now, now, 0, 0, request.getPriority())
               .setValues(ps);
 
           int insertedCount = ps.executeUpdate();