+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Describes notable changes.
 
+#### 1.21.1 - 2020/01/03
+- Task data is still saved into old data field, to allow seamless upgrade process.
+When upgrade from versions before 1.21.0 is finished, i.e. all cluster nodes have the new version;
+  `TasksProperties.Migration.copyDataToTwTaskField` can be set to `false`, stopping writing into the old data field.
+- `TasksProperties.Migration.previousVersion` is mandatory.
+
 #### 1.21.0 - 2020/12/21
 - Task data is now binary.
 - The payload is kept in the storage in compressed format.